The British have kidnapped and American and paraded her around as a sex slave "duchess"
The British have kidnapped and American and paraded her around as a sex slave "duchess". This is a declaration of war against America, that our population can be abducted and made to have slave children in another land. We must engage Prince Henry, while the family has removed him from home.